refactor(repository): ChannelRepository를 인터페이스로 만들고 JCFChannelRepository 클래스를 생성
- 원래 작성했던 ChannelRepository클래스를 JCFChannelRepository로 옮김
